refactor(skill-eval): readability cleanup from review-board pass 2
- Drop unnecessary generic from AffectedEntry and deduplicateAffectedEntries
- Clarify diffYamlEntries condition with comment distinguishing new vs previously non-skill entry
- Remove redundant length guards in collectSkillChanges (callees handle empty input)
- Move deduplicateAffectedEntries tests from skill-changes.test.ts to yaml.test.ts
